requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.11.12
cffi==2.0.0
hexdump==3.3
kafka-python==2.3.0
minio==7.2.20
numpy==2.3.5
pandas==2.3.3
pycparser==2.23
pycryptodome==3.23.0
python-dateutil==2.9.0.post0
python-evtx==0.8.1
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
watchdog==6.0.0

xmltodict
reportlab
ijson